scripts: Don't bail out when debugedit fails
In certain instances debugedit might just fail, and there is nothing we can
really do about it due to each individual package altering CFLAGS, LDFLAGs,
etc. Now we'll just gracefully continue to the next file in the list if
this happens, and leave the packager to continue with other issues
if needed.

Ported-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>
Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>